axiol/core/server/requirements.txt
cffi==1.14.6
fastapi==0.68.1
pycparser==2.20
pydantic==1.8.2
PyNaCl==1.4.0
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.2
